+++ release/utils/bindScrollPosition/bindScrollPosition.js
@@ -4,13 +4,11 @@ Object.defineProperty(exports, '__esModule', { value: true });
 
 function bindScrollPosition() {
     const listener = () => {
-        requestAnimationFrame(() => {
-            document.body.style.setProperty('--scroll', `${document.documentElement.scrollTop}px`);
-        });
+        document.body.style.setProperty('--scroll', `${window.scrollY}px`);
     };
-    document.addEventListener('scroll', listener);
+    window.addEventListener('scroll', listener, { capture: false, passive: true });
     listener();
-    return () => document.removeEventListener('scroll', listener);
+    return () => window.removeEventListener('scroll', listener);
 }
 
 exports.bindScrollPosition = bindScrollPosition;